test(orchestration): split focused contract coverage
